Add files via upload
These are the image targets uploaded to Vuforia that will be augmented by other images. They are the tasks-exercises that students should complete. They will be augmented by the images with the answers. The answers can be on top of the exercise, if it is a small one or on the next page for larger ones.